flink-job/requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
psycopg2-binary==2.9.9